+++ src/posix/mod.rs
@@ -4,6 +4,5 @@ pub use self::tty::*;
 mod enumerate;
 mod error;
 mod ioctl;
-mod poll;
 mod termios;
 mod tty;